+++ Nos/Service/RelayService.swift
@@ -29,6 +29,7 @@ final class RelayService: ObservableObject {
     private var parseQueue = ParseQueue()
     @Dependency(\.analytics) private var analytics
     @MainActor @Dependency(\.currentUser) private var currentUser
+    @Published var numberOfConnectedRelays: Int = 0
     
     init(persistenceController: PersistenceController) {
         self.persistenceController = persistenceController
@@ -179,6 +180,13 @@ extension RelayService {
         await clearStaleSubscriptions()
         
         await subscriptions.processSubscriptionQueue(relays: relays)
+        
+        let socketsCount = await subscriptions.sockets.count
+        Task { @MainActor in
+            if numberOfConnectedRelays != socketsCount {
+                numberOfConnectedRelays = socketsCount
+            }
+        }
     }
     
     private func clearStaleSubscriptions() async {